Only calculate sqrt if the dist is within bounds
We can compare the square of the distance (the value
before we take its root) to the square of each of the bounds
(one of which is zero), and then perform the sqrt operation
only if we are going to use the value.

Change comparison from the form 'x < a and a < y' to a chained
comparison of the form 'x < a < y'.

Use xrange instead of range.